refactor: presentation 내부 package 구조 변경
